+++ metadata/service/exporter/configurable/exporter.go
@@ -28,7 +28,6 @@ import (
 	"github.com/apache/dubbo-go/common/logger"
 	"github.com/apache/dubbo-go/config"
 	"github.com/apache/dubbo-go/metadata/service"
-	"github.com/apache/dubbo-go/metadata/service_exporter"
 )
 
 // MetadataServiceExporter is the ConfigurableMetadataServiceExporter which implement MetadataServiceExporter interface
@@ -39,7 +38,7 @@ type MetadataServiceExporter struct {
 }
 
 // NewMetadataServiceExporter will return a service_exporter.MetadataServiceExporter with the specified  metadata service
-func NewMetadataServiceExporter(metadataService service.MetadataService) service_exporter.MetadataServiceExporter {
+func NewMetadataServiceExporter(metadataService service.MetadataService) service.MetadataServiceExporter {
 	return &MetadataServiceExporter{
 		metadataService: metadataService,
 	}
