+++ metadata/service/exporter/configurable/exporter_test.go
@@ -18,6 +18,7 @@
 package configurable
 
 import (
+	"github.com/apache/dubbo-go/metadata/service"
 	"testing"
 )
 
@@ -54,7 +55,7 @@ func TestConfigurableExporter(t *testing.T) {
 		}})
 	config.MockInitProviderWithSingleRegistry()
 	metadataService := inmemory.NewMetadataService()
-	exported := NewMetadataServiceExporter(metadataService)
+	exported := service.NewMetadataServiceExporter(metadataService)
 	assert.Equal(t, false, exported.IsExported())
 	assert.NoError(t, exported.Export())
 	assert.Equal(t, true, exported.IsExported())
